requirements.txt
# Data Collection & Processing
icrawler==0.6.6
Pillow>=9.5.0
opencv-python>=4.8.0
albumentations>=1.3.1
tqdm>=4.65.0

# Deep Learning (YOLOv8)
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Backend & Deploy (Future)
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart
onnxruntime>=1.15.0
